+++ lib/src/fss_helpers.dart
@@ -1,5 +1,1 @@
 import 'package:ultimate_containers/ultimate_containers.dart';
-
-final shadow = FSS(
-  
-);