+++ lib/src/units/px.dart
@@ -12,10 +12,10 @@ class Px implements Unit {
   const Px(this.value);
 
   @override
-  CalculatedUnit operator +(Unit other) => CalculatedUnit.add(this, other);
+  Calc operator +(Unit other) => Calc.add(this, other);
 
   @override
-  CalculatedUnit operator -(Unit other) => CalculatedUnit.subtract(this, other);
+  Calc operator -(Unit other) => Calc.subtract(this, other);
 
   @override
   Px operator /(double factor) => Px(value * factor);