+++ lib/src/units/rem.dart
@@ -13,15 +13,15 @@ class Rem implements Unit {
   }
 
   @override
-  String toString() => '${value}px';
+  String toString() => '${value}rem';
 
   const Rem(this.value);
 
   @override
-  CalculatedUnit operator +(Unit other) => CalculatedUnit.add(this, other);
+  Calc operator +(Unit other) => Calc.add(this, other);
 
   @override
-  CalculatedUnit operator -(Unit other) => CalculatedUnit.subtract(this, other);
+  Calc operator -(Unit other) => Calc.subtract(this, other);
 
   @override
   Rem operator /(double factor) => Rem(value * factor);