+++ lib/ultimate_containers.dart
@@ -1,8 +1,11 @@
 library ultimate_containers;
 
 export 'src/components/components.dart';
-export 'src/fss.dart';
+export 'src/breakpoints_provider.dart';
+export 'src/breakpoints.dart';
+export 'src/fss_helpers.dart';
 export 'src/fss_provider.dart';
+export 'src/fss.dart';
 export 'src/units/units.dart';
 export 'src/unit_offset.dart';
 export 'src/unit_box_shadow.dart';