+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-## 1.0.0-dev1
+## 1.0.0-dev4
+
+- added `EC`component (`C` wrapped with `Expanded`)
+- added responsive design classes `Breakpoint` `Breakpoints` `BreakpointProvider` `BreakpointProviderBuilder`
+- added responsive sub styles `FSS.xs` `FSS.sm` `FSS.md` `FSS.lg` `FSS.xl` `FSS.xxl`
+- added functions for responsive caluclations `FSS.flattenResponsive` `FSS.getResponsive`
+- renamed `CaluclatedUnit` to `Calc`
+
+## 1.0.0-dev3
 
 - added borderWidth
 - added borderStyle